requirements.txt
certifi
requests
python-telegram-bot
transformers
torch
scipy
librosa
